change layout in Artifactory
The Dave Hammers script seems to like underscores.  Change the name
of the artifacts to use underscores to separate the parts.

Each distinct architecture needs its own subdirectory, since
the manifest cannot be used to specify the actual name of the
artifact.  So, 'subdir : telegraf/mips' would be used with
platform cougar, for example.

Tweak the ignore to match the new artifact names.