TELEGRAF-6: ah_wireless input plugins ioctl error
size of int is different in golang and C. In 32 bit
system, size of int is 4 in golang. But when 64 bit
system is used, size of int becomes 8. Where as in C
size of int is always 4 regardless of CPU. This
is creating struct interpretation wrong in ioctl
handler in 64 bit systems [AP5K]. Correcting it to
use int32 instead of int.